Change some columns name

+++ database/migrations/2020_02_07_131836_create_course_sublevels_table.php
@@ -15,7 +15,7 @@ class CreateCourseSublevelsTable extends Migration
     {
         Schema::create('course_sublevels', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('course_level_id');
+            $table->bigInteger('course_level_id');
             $table->string('title');
             $table->integer('time');
             $table->float('minimum_score');
