+++ database/migrations/2020_02_07_131846_create_course_levels_table.php
@@ -17,7 +17,7 @@ class CreateCourseLevelsTable extends Migration
             $table->bigIncrements('id');
             $table->timestamps();
             $table->string('title');
-            $table->integer('course_id');
+            $table->bigInteger('course_id');
         });
     }
